+++ src/types/mypage/language.type.ts
